+++ packages/moonstone/CHANGELOG.md
@@ -54,6 +54,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Notification` to align texts left in LTR locale and right in RTL locale
 - `moonstone/VideoPlayer` to simulate rewind functionality on non-webOS platforms only
 - `moonstone/Panels` to pass the `onTransition` prop to its underlying `<Viewport>`
+- `moonstone/Panels` to pass the `onTransition` and `onWillTransition` props to its underlying `<Viewport>`
 
 ### Fixed