+++ packages/moonstone/CHANGELOG.md
@@ -6,6 +6,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Deprecated
 
+- `moonstone/Input` prop `noDecorator` is being replaced by `autoFocus` in 2.0.0.
+
 ### Added
 
 ### Changed
@@ -33,7 +35,6 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/Scroller` props `horizontal` and `vertical`. Deprecated props are replaced with `direction` prop. `horizontal` and `vertical` will be removed in 2.0.0.
 - `moonstone/Panel` prop `noAutoFocus` in favor of `autoFocus="none"`
-- `moonstone/Input` prop `noDecorator` is being replaced by `autoFocus` in 2.0.0.
 
 ### Added